libcriterion: Fix default state display for InclusiveCriterion
InclusiveCriterion has a default state called "none", this state was
displayed through default Criterion behavior which is: if no valid state
is available, let's display "<none>".

This patch test explicitly the default value when formatting the state
of an InclusiveCriterion. It also introduces the test getLiteralValue
return when formatting a standard Criterion rather than testing the
generated string emptiness.
A difference can now be made between an inclusive criterion default
value and an exclusive criterion with an unknown state.

Signed-off-by: Jules Clero <julesx.clero@intel.com>