libcriterion: Rework state handling
Criterion API was allowing to set any state we want to any criterion.

This patch rework this API in order to add some check when setting
a criterion state.
It is now required that all parts of the state we set are registered
values.
Moreover inclusive criterion can now handle as many value as an
exclusive criterion (e.g as many values as an signed integer can
take).

Signed-off-by: Jules Clero <julesx.clero@intel.com>