libcriterion: Enhance default value handling
Criterion default value handling has been enhanced. Now, that
inclusive criterion is not represented by bit mask, 0 value can
be used as any other numerical value.

The "none" state of a criterion is now represented as an empty state.
This special state does not comes as any other standard, user provided
value, so it is not displayed as an available value.

Exclusive criterion default value handling remains the same, the first
value provided by the user is the default one. This value can be set
either explicitly or by using an empty state.

Signed-off-by: Jules Clero <julesx.clero@intel.com>