libcriterion: remove isInclusive API
Previously, client was sometimes needing to know either the criterion
was Inclusive or Exclusive.
Now that setState API has been secured, the client do not have to
realize any check. Moreover, if it really wants it, he can keep it after
the criterion creation.
This patch removes the isInclusive API.

Signed-off-by: Jules Clero <julesx.clero@intel.com>