libcriterion: Remove getFormattedState from public API
getFormattedState allows a client to retrieve properly a formatted
string containing all current state value in a formatted way.
Previous implementation was needing this helper as the state translation
to literal values was not straightforward.

Now that state is just a set of string, the client can easily display
this information and we don't need to impose him a format.
The method remains private as it is used for display purpose.

Signed-off-by: Jules Clero <julesx.clero@intel.com>